+++ raster/src/main/scala/geotrellis/raster/MultibandTile.scala
@@ -279,4 +279,8 @@ trait MultibandTile extends CellGrid with MacroCombinableMultibandTile[Tile] wit
     */
   def combineDouble(b0: Int, b1: Int)(f: (Double, Double) => Double): Tile
 
+  /**
+    * Convert the present [[MultibandTile]] to an [[MultibandTile]] of [[ArrayTile]]s.
+    */
+  def toArrayTile(): ArrayMultibandTile
 }