+++ raster/src/test/scala/geotrellis/raster/io/geotiff/SinglebandCropIteratorSpec.scala
@@ -67,10 +67,12 @@ class SinglebandCropIteratorSpec extends FunSpec
           geoTiff.raster.crop(256, 256, 512, 512))
 
       val actual: Array[Tile] =
-        Array(singlebandIterator.next.tile,
-          singlebandIterator.next.tile,
-          singlebandIterator.next.tile,
-          singlebandIterator.next.tile)
+        Array(
+          singlebandIterator.next.tile.toArrayTile,
+          singlebandIterator.next.tile.toArrayTile,
+          singlebandIterator.next.tile.toArrayTile,
+          singlebandIterator.next.tile.toArrayTile
+        )
 
       cfor(0)(_ < actual.length, _ + 1) { i =>
         assertEqual(expected(i), actual(i))
@@ -81,10 +83,10 @@ class SinglebandCropIteratorSpec extends FunSpec
       val windowedCols = 950
       val windowedRows = 1300
       val singlebandIterator =
-        new SinglebandCropIterator(geoTiff, windowedCols, windowedRows)
+        new SinglebandCropIterator(geoTiff.copy(tile = geoTiff.tile.toArrayTile), windowedCols, windowedRows)
 
-      val expected = geoTiff.tile
-      val actual = singlebandIterator.next.tile
+      val expected = geoTiff.tile.toArrayTile
+      val actual = singlebandIterator.next.tile.toArrayTile
 
       assertEqual(expected, actual)
     }
@@ -104,12 +106,14 @@ class SinglebandCropIteratorSpec extends FunSpec
           geoTiff.raster.crop(500, 450, 512, 512))
 
       val actual: Array[Tile] =
-        Array(singlebandIterator.next.tile,
-          singlebandIterator.next.tile,
-          singlebandIterator.next.tile,
-          singlebandIterator.next.tile,
-          singlebandIterator.next.tile,
-          singlebandIterator.next.tile)
+        Array(
+          singlebandIterator.next.tile.toArrayTile,
+          singlebandIterator.next.tile.toArrayTile,
+          singlebandIterator.next.tile.toArrayTile,
+          singlebandIterator.next.tile.toArrayTile,
+          singlebandIterator.next.tile.toArrayTile,
+          singlebandIterator.next.tile.toArrayTile
+        )
 
       cfor(0)(_ < actual.length, _ + 1) { i =>
         assertEqual(expected(i), actual(i))