+++ raster/src/test/scala/geotrellis/raster/io/geotiff/UByteGeoTiffMultibandTileSpec.scala
@@ -44,7 +44,7 @@ class ByteGeoTiffMultibandTileSpec extends FunSpec
 
     it("should combine all bands with pixel interleave, striped") {
       val tile =
-        MultibandGeoTiff.compressed(p("striped", "pixel")).tile
+        MultibandGeoTiff(p("striped", "pixel")).tile
 
       val actual = tile.combine(_.sum)
       val expected = UByteArrayTile(Array.ofDim[Byte](tile.cols * tile.rows).fill(6), tile.cols, tile.rows, UByteCellType)
@@ -54,7 +54,7 @@ class ByteGeoTiffMultibandTileSpec extends FunSpec
 
     it("should combine all bands with pixel interleave, tiled") {
       val tile =
-        MultibandGeoTiff.compressed(p("tiled", "pixel")).tile
+        MultibandGeoTiff(p("tiled", "pixel")).tile
 
       val actual = tile.combine(_.sum)
       val expected = UByteArrayTile(Array.ofDim[Byte](tile.cols * tile.rows).fill(6), tile.cols, tile.rows, UByteCellType)
@@ -64,7 +64,7 @@ class ByteGeoTiffMultibandTileSpec extends FunSpec
 
     it("should combine all bands with band interleave, striped") {
       val tile =
-        MultibandGeoTiff.compressed(p("striped", "band")).tile
+        MultibandGeoTiff(p("striped", "band")).tile
 
       val actual = tile.combine(_.sum)
       val expected = UByteArrayTile(Array.ofDim[Byte](tile.cols * tile.rows).fill(6), tile.cols, tile.rows, UByteCellType)
@@ -74,7 +74,7 @@ class ByteGeoTiffMultibandTileSpec extends FunSpec
 
     it("should combine all bands with band interleave, tiled") {
       val tile =
-        MultibandGeoTiff.compressed(p("tiled", "band")).tile
+        MultibandGeoTiff(p("tiled", "band")).tile
 
       val actual = tile.combine(_.sum)
       val expected = UByteArrayTile(Array.ofDim[Byte](tile.cols * tile.rows).fill(6), tile.cols, tile.rows, UByteCellType)
@@ -86,7 +86,7 @@ class ByteGeoTiffMultibandTileSpec extends FunSpec
 
     it("should combineDouble all bands with pixel interleave, striped") {
       val tile =
-        MultibandGeoTiff.compressed(p("striped", "pixel")).tile
+        MultibandGeoTiff(p("striped", "pixel")).tile
 
       val actual = tile.combineDouble(_.sum)
       val expected = UByteArrayTile(Array.ofDim[Byte](tile.cols * tile.rows).fill(6), tile.cols, tile.rows, UByteCellType)
@@ -96,7 +96,7 @@ class ByteGeoTiffMultibandTileSpec extends FunSpec
 
     it("should combineDouble all bands with pixel interleave, tiled") {
       val tile =
-        MultibandGeoTiff.compressed(p("tiled", "pixel")).tile
+        MultibandGeoTiff(p("tiled", "pixel")).tile
 
       val actual = tile.combineDouble(_.sum)
       val expected = UByteArrayTile(Array.ofDim[Byte](tile.cols * tile.rows).fill(6), tile.cols, tile.rows, UByteCellType)
@@ -106,7 +106,7 @@ class ByteGeoTiffMultibandTileSpec extends FunSpec
 
     it("should combineDouble all bands with band interleave, striped") {
       val tile =
-        MultibandGeoTiff.compressed(p("striped", "band")).tile
+        MultibandGeoTiff(p("striped", "band")).tile
 
       val actual = tile.combineDouble(_.sum)
       val expected = UByteArrayTile(Array.ofDim[Byte](tile.cols * tile.rows).fill(6), tile.cols, tile.rows, UByteCellType)
@@ -116,7 +116,7 @@ class ByteGeoTiffMultibandTileSpec extends FunSpec
 
     it("should combineDouble all bands with band interleave, tiled") {
       val tile =
-        MultibandGeoTiff.compressed(p("tiled", "band")).tile
+        MultibandGeoTiff(p("tiled", "band")).tile
 
       val actual = tile.combineDouble(_.sum)
       val expected = UByteArrayTile(Array.ofDim[Byte](tile.cols * tile.rows).fill(6), tile.cols, tile.rows, UByteCellType)
@@ -128,7 +128,7 @@ class ByteGeoTiffMultibandTileSpec extends FunSpec
 
     it("should combine 2 bands with pixel interleave, striped") {
       val tile =
-        MultibandGeoTiff.compressed(p("striped", "pixel")).tile
+        MultibandGeoTiff(p("striped", "pixel")).tile
 
       val actual = tile.combine(2, 1) { (z1, z2) => z1 + z2 }
       val expected = UByteArrayTile(Array.ofDim[Byte](tile.cols * tile.rows).fill(5), tile.cols, tile.rows, UByteCellType)
@@ -138,7 +138,7 @@ class ByteGeoTiffMultibandTileSpec extends FunSpec
 
     it("should combine 2 bands with pixel interleave, tiled") {
       val tile =
-        MultibandGeoTiff.compressed(p("tiled", "pixel")).tile
+        MultibandGeoTiff(p("tiled", "pixel")).tile
 
       val actual = tile.combine(1, 2) { (z1, z2) => z1 + z2 }
       val expected = UByteArrayTile(Array.ofDim[Byte](tile.cols * tile.rows).fill(5), tile.cols, tile.rows, UByteCellType)
@@ -148,7 +148,7 @@ class ByteGeoTiffMultibandTileSpec extends FunSpec
 
     it("should combine 2 bands with band interleave, striped") {
       val tile =
-        MultibandGeoTiff.compressed(p("striped", "band")).tile
+        MultibandGeoTiff(p("striped", "band")).tile
 
       val actual = tile.combine(1, 2) { (z1, z2) => z1 + z2 }
       val expected = UByteArrayTile(Array.ofDim[Byte](tile.cols * tile.rows).fill(5), tile.cols, tile.rows, UByteCellType)
@@ -158,7 +158,7 @@ class ByteGeoTiffMultibandTileSpec extends FunSpec
 
     it("should combine 2 bands with band interleave, tiled") {
       val tile =
-        MultibandGeoTiff.compressed(p("tiled", "band")).tile
+        MultibandGeoTiff(p("tiled", "band")).tile
 
       val actual = tile.combine(1, 2) { (z1, z2) => z1 + z2 }
       val expected = UByteArrayTile(Array.ofDim[Byte](tile.cols * tile.rows).fill(5), tile.cols, tile.rows, UByteCellType)
@@ -170,7 +170,7 @@ class ByteGeoTiffMultibandTileSpec extends FunSpec
 
     it("should combine 3 bands with pixel interleave, striped") {
       val tile =
-        MultibandGeoTiff.compressed(p("striped", "pixel")).tile
+        MultibandGeoTiff(p("striped", "pixel")).tile
 
       val actual = tile.combineDouble(2, 1, 0) { (z1, z2, z3) => z1 + z2 - z3 }
       val expected = UByteArrayTile(Array.ofDim[Byte](tile.cols * tile.rows).fill(4), tile.cols, tile.rows, UByteCellType)
@@ -180,7 +180,7 @@ class ByteGeoTiffMultibandTileSpec extends FunSpec
 
     it("should combine 3 bands with pixel interleave, tiled") {
       val tile =
-        MultibandGeoTiff.compressed(p("tiled", "pixel")).tile
+        MultibandGeoTiff(p("tiled", "pixel")).tile
 
       val actual = tile.combineDouble(0, 1, 2) { (z1, z2, z3) => z1 + z2 - z3 }
       val expected = UByteArrayTile(Array.ofDim[Byte](tile.cols * tile.rows).fill(0), tile.cols, tile.rows, UByteCellType)
@@ -190,7 +190,7 @@ class ByteGeoTiffMultibandTileSpec extends FunSpec
 
     it("should combine 3 bands with band interleave, striped") {
       val tile =
-        MultibandGeoTiff.compressed(p("striped", "band")).tile
+        MultibandGeoTiff(p("striped", "band")).tile
 
       val actual = tile.combineDouble(0, 1, 2) { (z1, z2, z3) => z1 * z2 - z3 }
       val expected = UByteArrayTile(Array.ofDim[Byte](tile.cols * tile.rows).fill(-1), tile.cols, tile.rows, UByteCellType)
@@ -200,7 +200,7 @@ class ByteGeoTiffMultibandTileSpec extends FunSpec
 
     it("should combine 3 bands with band interleave, tiled") {
       val tile =
-        MultibandGeoTiff.compressed(p("tiled", "band")).tile
+        MultibandGeoTiff(p("tiled", "band")).tile
 
       val actual = tile.combineDouble(0, 1, 2) { (z1, z2, z3) => z1 + (z2 * z3) }
       val expected = UByteArrayTile(Array.ofDim[Byte](tile.cols * tile.rows).fill(7), tile.cols, tile.rows, UByteCellType)
@@ -212,7 +212,7 @@ class ByteGeoTiffMultibandTileSpec extends FunSpec
 
     it("should combine 4 bands with pixel interleave, striped") {
       val tile =
-        MultibandGeoTiff.compressed(p("striped", "pixel")).tile
+        MultibandGeoTiff(p("striped", "pixel")).tile
 
       val actual = tile.combineDouble(2, 1, 0, 2) { (z1, z2, z3, z4) => z1 + z2 - z3 + z4}
       val expected = UByteArrayTile(Array.ofDim[Byte](tile.cols * tile.rows).fill(7), tile.cols, tile.rows, UByteCellType)
@@ -222,7 +222,7 @@ class ByteGeoTiffMultibandTileSpec extends FunSpec
 
     it("should combine 4 bands with pixel interleave, tiled") {
       val tile =
-        MultibandGeoTiff.compressed(p("tiled", "pixel")).tile
+        MultibandGeoTiff(p("tiled", "pixel")).tile
 
       val actual = tile.combineDouble(0, 1, 2, 0) { (z1, z2, z3, z4) => z1 + z2 - z3 + z4 }
       val expected = UByteArrayTile(Array.ofDim[Byte](tile.cols * tile.rows).fill(1), tile.cols, tile.rows, UByteCellType)
@@ -232,7 +232,7 @@ class ByteGeoTiffMultibandTileSpec extends FunSpec
 
     it("should combine 4 bands with band interleave, striped") {
       val tile =
-        MultibandGeoTiff.compressed(p("striped", "band")).tile
+        MultibandGeoTiff(p("striped", "band")).tile
 
       val actual = tile.combineDouble(0, 1, 2, 0) { (z1, z2, z3, z4) => z1 * z2 - z3 - z4}
       val expected = UByteArrayTile(Array.ofDim[Byte](tile.cols * tile.rows).fill(-2), tile.cols, tile.rows, UByteCellType)
@@ -242,7 +242,7 @@ class ByteGeoTiffMultibandTileSpec extends FunSpec
 
     it("should combine 4 bands with band interleave, tiled") {
       val tile =
-        MultibandGeoTiff.compressed(p("tiled", "band")).tile
+        MultibandGeoTiff(p("tiled", "band")).tile
 
       val actual = tile.combineDouble(0, 1, 2, 2) { (z1, z2, z3, z4) => z1 + (z2 * z3) + z4}
       val expected = UByteArrayTile(Array.ofDim[Byte](tile.cols * tile.rows).fill(10), tile.cols, tile.rows, UByteCellType)